# api/swagger/dist/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Education forum 2018 API",
    "version" : "1.0.0",
    "title" : "FORUM 2018 API",
    "contact" : {
      "email" : "vvovvrulit@gmail.com"
    }
  },
  "host" : "virtserver.swaggerhub.com",
  "basePath" : "/Forum/Forum/1.0.0",
...
  "schemes" : [ "https", "http" ],
  "paths" : {
...
    "/free-email" : {
      "post" : {
        "tags" : [ "/api/session" ],
        "summary" : "check if email is free",
        "description" : "check if email is free",
        "operationId" : "checkEmail",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "Email",
          "description" : "Email to check",
          "required" : false,
          "schema" : {
            "type" : "object",
            "example" : {
              "email" : "youremail@gmail.com"
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "email is free"
          },
          "400" : {
            "description" : "validation error"
          },
          "409" : {
            "description" : "db already exists this email"
          }
        }
      }
    },
...
  },
...
}